Fix login flow, and root loader validation check
